requirements.txt
flask
openai
google-generativeai
faiss-cpu
sentence-transformers
beautifulsoup4